// src/app/model/tipoNivel.ts
export interface TipoNivel {
  cdTipoNivel: number
  dsTipoDeNivel: string​
  flAtivo: boolean
  flExigeCentroDeCusto: boolean
  flPresidencia: boolean
  flTipoFormal: boolean​​
  nmAbreviatura: string
​​  nmTipoDeNivel: string
  modificationTime: string
  modifiedByUser: string
  creationTime: string
  createdByUser: string
}